Fixes and test for settings
