+++ morango/constants/transfer_stage.py
@@ -22,12 +22,13 @@ CHOICES = (
 )
 
 PRECEDENCE = {
-    SERIALIZING: 1,
-    QUEUING: 2,
-    TRANSFERRING: 3,
-    DEQUEUING: 4,
-    DESERIALIZING: 5,
-    CLEANUP: 6,
+    INITIALIZING: 1,
+    SERIALIZING: 2,
+    QUEUING: 3,
+    TRANSFERRING: 4,
+    DEQUEUING: 5,
+    DESERIALIZING: 6,
+    CLEANUP: 7,
 }
